rename dao entity class name
