+++ oauth-dal/src/main/java/org/zhenchao/oauth/dao/AppInfoMapper.java
@@ -1,0 +1,30 @@
+package org.zhenchao.oauth.dao;
+
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
+import org.zhenchao.oauth.entity.AppInfo;
+import org.zhenchao.oauth.entity.AppInfoExample;
+
+public interface AppInfoMapper {
+    long countByExample(AppInfoExample example);
+
+    int deleteByExample(AppInfoExample example);
+
+    int deleteByPrimaryKey(Long appId);
+
+    int insert(AppInfo record);
+
+    int insertSelective(AppInfo record);
+
+    List<AppInfo> selectByExample(AppInfoExample example);
+
+    AppInfo selectByPrimaryKey(Long appId);
+
+    int updateByExampleSelective(@Param("record") AppInfo record, @Param("example") AppInfoExample example);
+
+    int updateByExample(@Param("record") AppInfo record, @Param("example") AppInfoExample example);
+
+    int updateByPrimaryKeySelective(AppInfo record);
+
+    int updateByPrimaryKey(AppInfo record);
+}
