--- oauth-dal/src/main/java/org/zhenchao/oauth/dao/OAuthAppInfoMapper.java
@@ -1,97 +1,0 @@
-package org.zhenchao.oauth.dao;
-
-import org.apache.ibatis.annotations.Param;
-import org.zhenchao.oauth.model.OAuthAppInfo;
-import org.zhenchao.oauth.model.OAuthAppInfoExample;
-
-import java.util.List;
-
-public interface OAuthAppInfoMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    long countByExample(OAuthAppInfoExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int deleteByExample(OAuthAppInfoExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int deleteByPrimaryKey(Long appId);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int insert(OAuthAppInfo record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int insertSelective(OAuthAppInfo record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    List<OAuthAppInfo> selectByExample(OAuthAppInfoExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    OAuthAppInfo selectByPrimaryKey(Long appId);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int updateByExampleSelective(@Param("record") OAuthAppInfo record, @Param("example") OAuthAppInfoExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int updateByExample(@Param("record") OAuthAppInfo record, @Param("example") OAuthAppInfoExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int updateByPrimaryKeySelective(OAuthAppInfo record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    int updateByPrimaryKey(OAuthAppInfo record);
-}
