+++ oauth-dal/src/main/java/org/zhenchao/oauth/dao/ScopeMapper.java
@@ -1,97 +1,31 @@
 package org.zhenchao.oauth.dao;
 
 import org.apache.ibatis.annotations.Param;
-import org.zhenchao.oauth.model.Scope;
-import org.zhenchao.oauth.model.ScopeExample;
+import org.zhenchao.oauth.entity.Scope;
+import org.zhenchao.oauth.entity.ScopeExample;
 
 import java.util.List;
 
 public interface ScopeMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     long countByExample(ScopeExample example);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int deleteByExample(ScopeExample example);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int deleteByPrimaryKey(Integer id);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int insert(Scope record);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int insertSelective(Scope record);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     List<Scope> selectByExample(ScopeExample example);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     Scope selectByPrimaryKey(Integer id);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int updateByExampleSelective(@Param("record") Scope record, @Param("example") ScopeExample example);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int updateByExample(@Param("record") Scope record, @Param("example") ScopeExample example);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int updateByPrimaryKeySelective(Scope record);
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     int updateByPrimaryKey(Scope record);
 }
