+++ oauth-dal/src/main/java/org/zhenchao/oauth/dao/UserInfoMapper.java
@@ -1,0 +1,30 @@
+package org.zhenchao.oauth.dao;
+
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
+import org.zhenchao.oauth.entity.UserInfo;
+import org.zhenchao.oauth.entity.UserInfoExample;
+
+public interface UserInfoMapper {
+    long countByExample(UserInfoExample example);
+
+    int deleteByExample(UserInfoExample example);
+
+    int deleteByPrimaryKey(Long id);
+
+    int insert(UserInfo record);
+
+    int insertSelective(UserInfo record);
+
+    List<UserInfo> selectByExample(UserInfoExample example);
+
+    UserInfo selectByPrimaryKey(Long id);
+
+    int updateByExampleSelective(@Param("record") UserInfo record, @Param("example") UserInfoExample example);
+
+    int updateByExample(@Param("record") UserInfo record, @Param("example") UserInfoExample example);
+
+    int updateByPrimaryKeySelective(UserInfo record);
+
+    int updateByPrimaryKey(UserInfo record);
+}
