--- oauth-dal/src/main/java/org/zhenchao/oauth/dao/UserMapper.java
@@ -1,97 +1,0 @@
-package org.zhenchao.oauth.dao;
-
-import org.apache.ibatis.annotations.Param;
-import org.zhenchao.oauth.model.User;
-import org.zhenchao.oauth.model.UserExample;
-
-import java.util.List;
-
-public interface UserMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    long countByExample(UserExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int deleteByExample(UserExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int deleteByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int insert(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int insertSelective(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    List<User> selectByExample(UserExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    User selectByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int updateByExample(@Param("record") User record, @Param("example") UserExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int updateByPrimaryKeySelective(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    int updateByPrimaryKey(User record);
-}
