+++ oauth-dal/src/main/java/org/zhenchao/oauth/entity/AppInfo.java
@@ -1,0 +1,135 @@
+package org.zhenchao.oauth.entity;
+
+import java.util.Date;
+
+public class AppInfo {
+    private Long appId;
+
+    private String appName;
+
+    private String logo;
+
+    private Boolean enable;
+
+    private String redirectUri;
+
+    private String cancelRedirectUri;
+
+    private String scope;
+
+    private Integer tokenValidity;
+
+    private String secret;
+
+    private Integer level;
+
+    private Long creatorId;
+
+    private Date createTime;
+
+    private Date updateTime;
+
+    public Long getAppId() {
+        return appId;
+    }
+
+    public void setAppId(Long appId) {
+        this.appId = appId;
+    }
+
+    public String getAppName() {
+        return appName;
+    }
+
+    public void setAppName(String appName) {
+        this.appName = appName == null ? null : appName.trim();
+    }
+
+    public String getLogo() {
+        return logo;
+    }
+
+    public void setLogo(String logo) {
+        this.logo = logo == null ? null : logo.trim();
+    }
+
+    public Boolean getEnable() {
+        return enable;
+    }
+
+    public void setEnable(Boolean enable) {
+        this.enable = enable;
+    }
+
+    public String getRedirectUri() {
+        return redirectUri;
+    }
+
+    public void setRedirectUri(String redirectUri) {
+        this.redirectUri = redirectUri == null ? null : redirectUri.trim();
+    }
+
+    public String getCancelRedirectUri() {
+        return cancelRedirectUri;
+    }
+
+    public void setCancelRedirectUri(String cancelRedirectUri) {
+        this.cancelRedirectUri = cancelRedirectUri == null ? null : cancelRedirectUri.trim();
+    }
+
+    public String getScope() {
+        return scope;
+    }
+
+    public void setScope(String scope) {
+        this.scope = scope == null ? null : scope.trim();
+    }
+
+    public Integer getTokenValidity() {
+        return tokenValidity;
+    }
+
+    public void setTokenValidity(Integer tokenValidity) {
+        this.tokenValidity = tokenValidity;
+    }
+
+    public String getSecret() {
+        return secret;
+    }
+
+    public void setSecret(String secret) {
+        this.secret = secret == null ? null : secret.trim();
+    }
+
+    public Integer getLevel() {
+        return level;
+    }
+
+    public void setLevel(Integer level) {
+        this.level = level;
+    }
+
+    public Long getCreatorId() {
+        return creatorId;
+    }
+
+    public void setCreatorId(Long creatorId) {
+        this.creatorId = creatorId;
+    }
+
+    public Date getCreateTime() {
+        return createTime;
+    }
+
+    public void setCreateTime(Date createTime) {
+        this.createTime = createTime;
+    }
+
+    public Date getUpdateTime() {
+        return updateTime;
+    }
+
+    public void setUpdateTime(Date updateTime) {
+        this.updateTime = updateTime;
+    }
+}
