+++ oauth-dal/src/main/java/org/zhenchao/oauth/entity/AppInfoExample.java
@@ -1,122 +1,50 @@
-package org.zhenchao.oauth.model;
+package org.zhenchao.oauth.entity;
 
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class OAuthAppInfoExample {
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
+public class AppInfoExample {
     protected String orderByClause;
 
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     protected boolean distinct;
 
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     protected List<Criteria> oredCriteria;
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public OAuthAppInfoExample() {
+    public AppInfoExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public void setOrderByClause(String orderByClause) {
         this.orderByClause = orderByClause;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public String getOrderByClause() {
         return orderByClause;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public void setDistinct(boolean distinct) {
         this.distinct = distinct;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public boolean isDistinct() {
         return distinct;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public List<Criteria> getOredCriteria() {
         return oredCriteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public void or(Criteria criteria) {
         oredCriteria.add(criteria);
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public Criteria or() {
         Criteria criteria = createCriteriaInternal();
         oredCriteria.add(criteria);
         return criteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public Criteria createCriteria() {
         Criteria criteria = createCriteriaInternal();
         if (oredCriteria.size() == 0) {
@@ -125,35 +53,17 @@ public class OAuthAppInfoExample {
         return criteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     protected Criteria createCriteriaInternal() {
         Criteria criteria = new Criteria();
         return criteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public void clear() {
         oredCriteria.clear();
         orderByClause = null;
         distinct = false;
     }
 
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     protected abstract static class GeneratedCriteria {
         protected List<Criterion> criteria;
 
@@ -1036,12 +946,6 @@ public class OAuthAppInfoExample {
         }
     }
 
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated do_not_delete_during_merge Sat Feb 18 14:39:44 CST 2017
-     */
     public static class Criteria extends GeneratedCriteria {
 
         protected Criteria() {
@@ -1049,12 +953,6 @@ public class OAuthAppInfoExample {
         }
     }
 
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_oauth_app_info
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
     public static class Criterion {
         private String condition;
