+++ oauth-dal/src/main/java/org/zhenchao/oauth/entity/AuthorizeRelation.java
@@ -1,0 +1,95 @@
+package org.zhenchao.oauth.entity;
+
+import java.util.Date;
+
+public class AuthorizeRelation {
+    private Long appId;
+
+    private Long userId;
+
+    private String scope;
+
+    private String scopeSign;
+
+    private String tokenKey;
+
+    private String refreshTokenKey;
+
+    private Long refreshTokenExpirationTime;
+
+    private Date createTime;
+
+    private Date updateTime;
+
+    public Long getAppId() {
+        return appId;
+    }
+
+    public void setAppId(Long appId) {
+        this.appId = appId;
+    }
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+
+    public String getScope() {
+        return scope;
+    }
+
+    public void setScope(String scope) {
+        this.scope = scope == null ? null : scope.trim();
+    }
+
+    public String getScopeSign() {
+        return scopeSign;
+    }
+
+    public void setScopeSign(String scopeSign) {
+        this.scopeSign = scopeSign == null ? null : scopeSign.trim();
+    }
+
+    public String getTokenKey() {
+        return tokenKey;
+    }
+
+    public void setTokenKey(String tokenKey) {
+        this.tokenKey = tokenKey == null ? null : tokenKey.trim();
+    }
+
+    public String getRefreshTokenKey() {
+        return refreshTokenKey;
+    }
+
+    public void setRefreshTokenKey(String refreshTokenKey) {
+        this.refreshTokenKey = refreshTokenKey == null ? null : refreshTokenKey.trim();
+    }
+
+    public Long getRefreshTokenExpirationTime() {
+        return refreshTokenExpirationTime;
+    }
+
+    public void setRefreshTokenExpirationTime(Long refreshTokenExpirationTime) {
+        this.refreshTokenExpirationTime = refreshTokenExpirationTime;
+    }
+
+    public Date getCreateTime() {
+        return createTime;
+    }
+
+    public void setCreateTime(Date createTime) {
+        this.createTime = createTime;
+    }
+
+    public Date getUpdateTime() {
+        return updateTime;
+    }
+
+    public void setUpdateTime(Date updateTime) {
+        this.updateTime = updateTime;
+    }
+}
