+++ oauth-dal/src/main/java/org/zhenchao/oauth/entity/AuthorizeRelationExample.java
@@ -1,122 +1,50 @@
-package org.zhenchao.oauth.model;
+package org.zhenchao.oauth.entity;
 
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class UserAppAuthorizationExample {
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
+public class AuthorizeRelationExample {
     protected String orderByClause;
 
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     protected boolean distinct;
 
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     protected List<Criteria> oredCriteria;
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public UserAppAuthorizationExample() {
+    public AuthorizeRelationExample() {
         oredCriteria = new ArrayList<Criteria>();
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public void setOrderByClause(String orderByClause) {
         this.orderByClause = orderByClause;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public String getOrderByClause() {
         return orderByClause;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public void setDistinct(boolean distinct) {
         this.distinct = distinct;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public boolean isDistinct() {
         return distinct;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public List<Criteria> getOredCriteria() {
         return oredCriteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public void or(Criteria criteria) {
         oredCriteria.add(criteria);
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public Criteria or() {
         Criteria criteria = createCriteriaInternal();
         oredCriteria.add(criteria);
         return criteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public Criteria createCriteria() {
         Criteria criteria = createCriteriaInternal();
         if (oredCriteria.size() == 0) {
@@ -125,35 +53,17 @@ public class UserAppAuthorizationExample {
         return criteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     protected Criteria createCriteriaInternal() {
         Criteria criteria = new Criteria();
         return criteria;
     }
 
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public void clear() {
         oredCriteria.clear();
         orderByClause = null;
         distinct = false;
     }
 
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     protected abstract static class GeneratedCriteria {
         protected List<Criterion> criteria;
 
@@ -776,12 +686,6 @@ public class UserAppAuthorizationExample {
         }
     }
 
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated do_not_delete_during_merge Wed Dec 28 16:41:47 CST 2016
-     */
     public static class Criteria extends GeneratedCriteria {
 
         protected Criteria() {
@@ -789,12 +693,6 @@ public class UserAppAuthorizationExample {
         }
     }
 
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table t_user_app_authorization
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
     public static class Criterion {
         private String condition;
