+++ oauth-dal/src/main/java/org/zhenchao/oauth/mapper/ScopeMapper.xml
@@ -1,12 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="org.zhenchao.oauth.dao.ScopeMapper">
-  <resultMap id="BaseResultMap" type="org.zhenchao.oauth.model.Scope">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <resultMap id="BaseResultMap" type="org.zhenchao.oauth.entity.Scope">
     <id column="id" jdbcType="INTEGER" property="id" />
     <result column="name" jdbcType="VARCHAR" property="name" />
     <result column="description" jdbcType="VARCHAR" property="description" />
@@ -14,11 +9,6 @@
     <result column="type" jdbcType="INTEGER" property="type" />
   </resultMap>
   <sql id="Example_Where_Clause">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     <where>
       <foreach collection="oredCriteria" item="criteria" separator="or">
         <if test="criteria.valid">
@@ -48,11 +38,6 @@
     </where>
   </sql>
   <sql id="Update_By_Example_Where_Clause">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     <where>
       <foreach collection="example.oredCriteria" item="criteria" separator="or">
         <if test="criteria.valid">
@@ -82,19 +67,9 @@
     </where>
   </sql>
   <sql id="Base_Column_List">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     id, name, description, level, type
   </sql>
-  <select id="selectByExample" parameterType="org.zhenchao.oauth.model.ScopeExample" resultMap="BaseResultMap">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <select id="selectByExample" parameterType="org.zhenchao.oauth.entity.ScopeExample" resultMap="BaseResultMap">
     select
     <if test="distinct">
       distinct
@@ -109,53 +84,28 @@
     </if>
   </select>
   <select id="selectByPrimaryKey" parameterType="java.lang.Integer" resultMap="BaseResultMap">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     select 
     <include refid="Base_Column_List" />
     from t_scope
     where id = #{id,jdbcType=INTEGER}
   </select>
   <delete id="deleteByPrimaryKey" parameterType="java.lang.Integer">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     delete from t_scope
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="org.zhenchao.oauth.model.ScopeExample">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <delete id="deleteByExample" parameterType="org.zhenchao.oauth.entity.ScopeExample">
     delete from t_scope
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="org.zhenchao.oauth.model.Scope">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <insert id="insert" parameterType="org.zhenchao.oauth.entity.Scope">
     insert into t_scope (id, name, description, 
       level, type)
     values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{description,jdbcType=VARCHAR}, 
       #{level,jdbcType=INTEGER}, #{type,jdbcType=INTEGER})
   </insert>
-  <insert id="insertSelective" parameterType="org.zhenchao.oauth.model.Scope">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <insert id="insertSelective" parameterType="org.zhenchao.oauth.entity.Scope">
     insert into t_scope
     <trim prefix="(" suffix=")" suffixOverrides=",">
       <if test="id != null">
@@ -192,23 +142,13 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="org.zhenchao.oauth.model.ScopeExample" resultType="java.lang.Long">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <select id="countByExample" parameterType="org.zhenchao.oauth.entity.ScopeExample" resultType="java.lang.Long">
     select count(*) from t_scope
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
     </if>
   </select>
   <update id="updateByExampleSelective" parameterType="map">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     update t_scope
     <set>
       <if test="record.id != null">
@@ -232,11 +172,6 @@
     </if>
   </update>
   <update id="updateByExample" parameterType="map">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
     update t_scope
     set id = #{record.id,jdbcType=INTEGER},
       name = #{record.name,jdbcType=VARCHAR},
@@ -247,12 +182,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="org.zhenchao.oauth.model.Scope">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <update id="updateByPrimaryKeySelective" parameterType="org.zhenchao.oauth.entity.Scope">
     update t_scope
     <set>
       <if test="name != null">
@@ -270,12 +200,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="org.zhenchao.oauth.model.Scope">
-    <!--
-      WARNING - @mbg.generated
-      This element is automatically generated by MyBatis Generator, do not modify.
-      This element was generated on Wed Dec 28 16:41:47 CST 2016.
-    -->
+  <update id="updateByPrimaryKey" parameterType="org.zhenchao.oauth.entity.Scope">
     update t_scope
     set name = #{name,jdbcType=VARCHAR},
       description = #{description,jdbcType=VARCHAR},
