--- oauth-dal/src/main/java/org/zhenchao/oauth/model/OAuthAppInfo.java
@@ -1,442 +1,0 @@
-package org.zhenchao.oauth.model;
-
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-import java.util.Date;
-
-public class OAuthAppInfo {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.app_id
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Long appId;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.app_name
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private String appName;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.logo
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private String logo;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.enable
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Boolean enable;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.redirect_uri
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private String redirectUri;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.cancel_redirect_uri
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private String cancelRedirectUri;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.scope
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private String scope;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.token_validity
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Integer tokenValidity;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.secret
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private String secret;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.level
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Integer level;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.creator_id
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Long creatorId;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.create_time
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Date createTime;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_oauth_app_info.update_time
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    private Date updateTime;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.app_id
-     *
-     * @return the value of t_oauth_app_info.app_id
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Long getAppId() {
-        return appId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.app_id
-     *
-     * @param appId the value for t_oauth_app_info.app_id
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setAppId(Long appId) {
-        this.appId = appId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.app_name
-     *
-     * @return the value of t_oauth_app_info.app_name
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public String getAppName() {
-        return appName;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.app_name
-     *
-     * @param appName the value for t_oauth_app_info.app_name
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setAppName(String appName) {
-        this.appName = appName == null ? null : appName.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.logo
-     *
-     * @return the value of t_oauth_app_info.logo
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public String getLogo() {
-        return logo;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.logo
-     *
-     * @param logo the value for t_oauth_app_info.logo
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setLogo(String logo) {
-        this.logo = logo == null ? null : logo.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.enable
-     *
-     * @return the value of t_oauth_app_info.enable
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Boolean getEnable() {
-        return enable;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.enable
-     *
-     * @param enable the value for t_oauth_app_info.enable
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setEnable(Boolean enable) {
-        this.enable = enable;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.redirect_uri
-     *
-     * @return the value of t_oauth_app_info.redirect_uri
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public String getRedirectUri() {
-        return redirectUri;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.redirect_uri
-     *
-     * @param redirectUri the value for t_oauth_app_info.redirect_uri
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setRedirectUri(String redirectUri) {
-        this.redirectUri = redirectUri == null ? null : redirectUri.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.cancel_redirect_uri
-     *
-     * @return the value of t_oauth_app_info.cancel_redirect_uri
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public String getCancelRedirectUri() {
-        return cancelRedirectUri;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.cancel_redirect_uri
-     *
-     * @param cancelRedirectUri the value for t_oauth_app_info.cancel_redirect_uri
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setCancelRedirectUri(String cancelRedirectUri) {
-        this.cancelRedirectUri = cancelRedirectUri == null ? null : cancelRedirectUri.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.scope
-     *
-     * @return the value of t_oauth_app_info.scope
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public String getScope() {
-        return scope;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.scope
-     *
-     * @param scope the value for t_oauth_app_info.scope
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setScope(String scope) {
-        this.scope = scope == null ? null : scope.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.token_validity
-     *
-     * @return the value of t_oauth_app_info.token_validity
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Integer getTokenValidity() {
-        return tokenValidity;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.token_validity
-     *
-     * @param tokenValidity the value for t_oauth_app_info.token_validity
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setTokenValidity(Integer tokenValidity) {
-        this.tokenValidity = tokenValidity;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.secret
-     *
-     * @return the value of t_oauth_app_info.secret
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public String getSecret() {
-        return secret;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.secret
-     *
-     * @param secret the value for t_oauth_app_info.secret
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setSecret(String secret) {
-        this.secret = secret == null ? null : secret.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.level
-     *
-     * @return the value of t_oauth_app_info.level
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Integer getLevel() {
-        return level;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.level
-     *
-     * @param level the value for t_oauth_app_info.level
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setLevel(Integer level) {
-        this.level = level;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.creator_id
-     *
-     * @return the value of t_oauth_app_info.creator_id
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Long getCreatorId() {
-        return creatorId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.creator_id
-     *
-     * @param creatorId the value for t_oauth_app_info.creator_id
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setCreatorId(Long creatorId) {
-        this.creatorId = creatorId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.create_time
-     *
-     * @return the value of t_oauth_app_info.create_time
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Date getCreateTime() {
-        return createTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.create_time
-     *
-     * @param createTime the value for t_oauth_app_info.create_time
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_oauth_app_info.update_time
-     *
-     * @return the value of t_oauth_app_info.update_time
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public Date getUpdateTime() {
-        return updateTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_oauth_app_info.update_time
-     *
-     * @param updateTime the value for t_oauth_app_info.update_time
-     *
-     * @mbg.generated Sat Feb 18 14:39:44 CST 2017
-     */
-    public void setUpdateTime(Date updateTime) {
-        this.updateTime = updateTime;
-    }
-
-    @Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this, ToStringStyle.JSON_STYLE);
-    }
-}
