--- oauth-dal/src/main/java/org/zhenchao/oauth/model/Scope.java
@@ -1,168 +1,0 @@
-package org.zhenchao.oauth.model;
-
-public class Scope {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_scope.id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Integer id;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_scope.name
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String name;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_scope.description
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String description;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_scope.level
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Integer level;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_scope.type
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Integer type;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_scope.id
-     *
-     * @return the value of t_scope.id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Integer getId() {
-        return id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_scope.id
-     *
-     * @param id the value for t_scope.id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_scope.name
-     *
-     * @return the value of t_scope.name
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_scope.name
-     *
-     * @param name the value for t_scope.name
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setName(String name) {
-        this.name = name == null ? null : name.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_scope.description
-     *
-     * @return the value of t_scope.description
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getDescription() {
-        return description;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_scope.description
-     *
-     * @param description the value for t_scope.description
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setDescription(String description) {
-        this.description = description == null ? null : description.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_scope.level
-     *
-     * @return the value of t_scope.level
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Integer getLevel() {
-        return level;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_scope.level
-     *
-     * @param level the value for t_scope.level
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setLevel(Integer level) {
-        this.level = level;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_scope.type
-     *
-     * @return the value of t_scope.type
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Integer getType() {
-        return type;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_scope.type
-     *
-     * @param type the value for t_scope.type
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setType(Integer type) {
-        this.type = type;
-    }
-}
