--- oauth-dal/src/main/java/org/zhenchao/oauth/model/User.java
@@ -1,234 +1,0 @@
-package org.zhenchao.oauth.model;
-
-public class User {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Long id;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.username
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String username;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.password
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String password;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.age
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Integer age;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.phone
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String phone;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.email
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String email;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.avatar
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String avatar;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.id
-     *
-     * @return the value of t_user.id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Long getId() {
-        return id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.id
-     *
-     * @param id the value for t_user.id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.username
-     *
-     * @return the value of t_user.username
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getUsername() {
-        return username;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.username
-     *
-     * @param username the value for t_user.username
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setUsername(String username) {
-        this.username = username == null ? null : username.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.password
-     *
-     * @return the value of t_user.password
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getPassword() {
-        return password;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.password
-     *
-     * @param password the value for t_user.password
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setPassword(String password) {
-        this.password = password == null ? null : password.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.age
-     *
-     * @return the value of t_user.age
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Integer getAge() {
-        return age;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.age
-     *
-     * @param age the value for t_user.age
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.phone
-     *
-     * @return the value of t_user.phone
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getPhone() {
-        return phone;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.phone
-     *
-     * @param phone the value for t_user.phone
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setPhone(String phone) {
-        this.phone = phone == null ? null : phone.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.email
-     *
-     * @return the value of t_user.email
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getEmail() {
-        return email;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.email
-     *
-     * @param email the value for t_user.email
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setEmail(String email) {
-        this.email = email == null ? null : email.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.avatar
-     *
-     * @return the value of t_user.avatar
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getAvatar() {
-        return avatar;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.avatar
-     *
-     * @param avatar the value for t_user.avatar
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setAvatar(String avatar) {
-        this.avatar = avatar == null ? null : avatar.trim();
-    }
-}
