--- oauth-dal/src/main/java/org/zhenchao/oauth/model/UserAppAuthorization.java
@@ -1,302 +1,0 @@
-package org.zhenchao.oauth.model;
-
-import java.util.Date;
-
-public class UserAppAuthorization {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.app_id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Long appId;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.user_id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Long userId;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String scope;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.scope_sign
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String scopeSign;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.token_key
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String tokenKey;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.refresh_token_key
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private String refreshTokenKey;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.refresh_token_expiration_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Long refreshTokenExpirationTime;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.create_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Date createTime;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user_app_authorization.update_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    private Date updateTime;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.app_id
-     *
-     * @return the value of t_user_app_authorization.app_id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Long getAppId() {
-        return appId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.app_id
-     *
-     * @param appId the value for t_user_app_authorization.app_id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setAppId(Long appId) {
-        this.appId = appId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.user_id
-     *
-     * @return the value of t_user_app_authorization.user_id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Long getUserId() {
-        return userId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.user_id
-     *
-     * @param userId the value for t_user_app_authorization.user_id
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.scope
-     *
-     * @return the value of t_user_app_authorization.scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getScope() {
-        return scope;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.scope
-     *
-     * @param scope the value for t_user_app_authorization.scope
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setScope(String scope) {
-        this.scope = scope == null ? null : scope.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.scope_sign
-     *
-     * @return the value of t_user_app_authorization.scope_sign
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getScopeSign() {
-        return scopeSign;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.scope_sign
-     *
-     * @param scopeSign the value for t_user_app_authorization.scope_sign
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setScopeSign(String scopeSign) {
-        this.scopeSign = scopeSign == null ? null : scopeSign.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.token_key
-     *
-     * @return the value of t_user_app_authorization.token_key
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getTokenKey() {
-        return tokenKey;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.token_key
-     *
-     * @param tokenKey the value for t_user_app_authorization.token_key
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setTokenKey(String tokenKey) {
-        this.tokenKey = tokenKey == null ? null : tokenKey.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.refresh_token_key
-     *
-     * @return the value of t_user_app_authorization.refresh_token_key
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public String getRefreshTokenKey() {
-        return refreshTokenKey;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.refresh_token_key
-     *
-     * @param refreshTokenKey the value for t_user_app_authorization.refresh_token_key
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setRefreshTokenKey(String refreshTokenKey) {
-        this.refreshTokenKey = refreshTokenKey == null ? null : refreshTokenKey.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.refresh_token_expiration_time
-     *
-     * @return the value of t_user_app_authorization.refresh_token_expiration_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Long getRefreshTokenExpirationTime() {
-        return refreshTokenExpirationTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.refresh_token_expiration_time
-     *
-     * @param refreshTokenExpirationTime the value for t_user_app_authorization.refresh_token_expiration_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setRefreshTokenExpirationTime(Long refreshTokenExpirationTime) {
-        this.refreshTokenExpirationTime = refreshTokenExpirationTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.create_time
-     *
-     * @return the value of t_user_app_authorization.create_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Date getCreateTime() {
-        return createTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.create_time
-     *
-     * @param createTime the value for t_user_app_authorization.create_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user_app_authorization.update_time
-     *
-     * @return the value of t_user_app_authorization.update_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public Date getUpdateTime() {
-        return updateTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user_app_authorization.update_time
-     *
-     * @param updateTime the value for t_user_app_authorization.update_time
-     *
-     * @mbg.generated Wed Dec 28 16:41:47 CST 2016
-     */
-    public void setUpdateTime(Date updateTime) {
-        this.updateTime = updateTime;
-    }
-}
