+++ oauth-service/src/main/java/org/zhenchao/oauth/service/impl/AuthorizeRelationServiceImpl.java
@@ -5,9 +5,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import org.zhenchao.oauth.dao.UserAppAuthorizationMapper;
-import org.zhenchao.oauth.model.UserAppAuthorization;
-import org.zhenchao.oauth.model.UserAppAuthorizationExample;
+import org.zhenchao.oauth.dao.AuthorizeRelationMapper;
+import org.zhenchao.oauth.entity.AuthorizeRelation;
+import org.zhenchao.oauth.entity.AuthorizeRelationExample;
 import org.zhenchao.oauth.service.AuthorizeRelationService;
 
 import java.util.ArrayList;
@@ -27,30 +27,30 @@ public class AuthorizeRelationServiceImpl implements AuthorizeRelationService {
     private static final Logger log = LoggerFactory.getLogger(AuthorizeRelationServiceImpl.class);
 
     @Resource
-    private UserAppAuthorizationMapper userAppAuthorizationMapper;
+    private AuthorizeRelationMapper authorizeRelationMapper;
 
     @Override
-    public List<UserAppAuthorization> getUserAndAppAuthorizationInfo(long userId, long appId) {
+    public List<AuthorizeRelation> getUserAndAppRelationList(long userId, long appId) {
         if (userId < 0 || appId < 0) {
             log.error("Get user and app authorization error, userId[{}] or appId[{}] less than 0!", userId, appId);
             return new ArrayList<>();
         }
-        UserAppAuthorizationExample example = new UserAppAuthorizationExample();
-        UserAppAuthorizationExample.Criteria criteria = example.createCriteria();
+        AuthorizeRelationExample example = new AuthorizeRelationExample();
+        AuthorizeRelationExample.Criteria criteria = example.createCriteria();
         criteria.andAppIdEqualTo(appId).andUserIdEqualTo(userId);
-        return userAppAuthorizationMapper.selectByExample(example);
+        return authorizeRelationMapper.selectByExample(example);
     }
 
     @Override
-    public Optional<UserAppAuthorization> getUserAndAppAuthorizationInfo(long userId, long appId, String scopeSign) {
+    public Optional<AuthorizeRelation> getUserAndAppRelationList(long userId, long appId, String scopeSign) {
         if (userId < 0 || appId < 0 || StringUtils.isBlank(scopeSign)) {
             log.error("Get user and app authorization params error, userId[{}], appId[{}], scope sign [{}]!", userId, appId, scopeSign);
             return Optional.empty();
         }
-        UserAppAuthorizationExample example = new UserAppAuthorizationExample();
-        UserAppAuthorizationExample.Criteria criteria = example.createCriteria();
+        AuthorizeRelationExample example = new AuthorizeRelationExample();
+        AuthorizeRelationExample.Criteria criteria = example.createCriteria();
         criteria.andAppIdEqualTo(appId).andUserIdEqualTo(userId).andScopeSignEqualTo(scopeSign);
-        List<UserAppAuthorization> authorizations = userAppAuthorizationMapper.selectByExample(example);
+        List<AuthorizeRelation> authorizations = authorizeRelationMapper.selectByExample(example);
         if (CollectionUtils.isEmpty(authorizations)) {
             return Optional.empty();
         }
@@ -58,30 +58,30 @@ public class AuthorizeRelationServiceImpl implements AuthorizeRelationService {
     }
 
     @Override
-    public boolean replaceUserAndAppAuthorizationInfo(UserAppAuthorization authorization) {
-        if (null == authorization || StringUtils.isBlank(authorization.getScopeSign())) {
+    public boolean replaceAuthorizeRelation(AuthorizeRelation relation) {
+        if (null == relation || StringUtils.isBlank(relation.getScopeSign())) {
             return false;
         }
-        UserAppAuthorizationExample example = new UserAppAuthorizationExample();
-        UserAppAuthorizationExample.Criteria criteria = example.createCriteria();
+        AuthorizeRelationExample example = new AuthorizeRelationExample();
+        AuthorizeRelationExample.Criteria criteria = example.createCriteria();
         criteria.andUserIdEqualTo(
-                authorization.getUserId()).andAppIdEqualTo(authorization.getAppId()).andScopeSignEqualTo(authorization.getScopeSign());
-        List<UserAppAuthorization> authorizations = userAppAuthorizationMapper.selectByExample(example);
+                relation.getUserId()).andAppIdEqualTo(relation.getAppId()).andScopeSignEqualTo(relation.getScopeSign());
+        List<AuthorizeRelation> authorizations = authorizeRelationMapper.selectByExample(example);
 
         if (CollectionUtils.isEmpty(authorizations)) {
             // 用户与APP之间不存在特定授权关系
-            return userAppAuthorizationMapper.insertSelective(authorization) == 1;
+            return authorizeRelationMapper.insertSelective(relation) == 1;
         } else {
             // 用户与APP之间存在特定授权关系
-            UserAppAuthorization uaa = authorizations.get(0);
-            uaa.setRefreshTokenExpirationTime(authorization.getRefreshTokenExpirationTime());
-            uaa.setRefreshTokenKey(authorization.getRefreshTokenKey());
-            uaa.setTokenKey(authorization.getTokenKey());
-            uaa.setUpdateTime(authorization.getUpdateTime());
-            example = new UserAppAuthorizationExample();
+            AuthorizeRelation ar = authorizations.get(0);
+            ar.setRefreshTokenExpirationTime(relation.getRefreshTokenExpirationTime());
+            ar.setRefreshTokenKey(relation.getRefreshTokenKey());
+            ar.setTokenKey(relation.getTokenKey());
+            ar.setUpdateTime(relation.getUpdateTime());
+            example = new AuthorizeRelationExample();
             criteria = example.createCriteria();
-            criteria.andUserIdEqualTo(uaa.getUserId()).andAppIdEqualTo(uaa.getAppId()).andScopeSignEqualTo(authorization.getScopeSign());
-            return userAppAuthorizationMapper.updateByExampleSelective(uaa, example) == 1;
+            criteria.andUserIdEqualTo(ar.getUserId()).andAppIdEqualTo(ar.getAppId()).andScopeSignEqualTo(relation.getScopeSign());
+            return authorizeRelationMapper.updateByExampleSelective(ar, example) == 1;
         }
     }
